Cleaned up Libft

+++ ft_strmapi.c
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	unsigned int	i;
-	char			*str;
+	size_t	i;
+	char	*str;
 
 	i = 0;
 	str = malloc(sizeof(char) * (ft_strlen(s) + 1));
